+++ lib/ndr_dev_support/capistrano/svn_cache.rb
@@ -17,7 +17,13 @@ Capistrano::Configuration.instance(:must_exist).load do
         system("rm -rf #{cache}")
       end
     end
+
+    desc 'Ensures compilation artefacts are removed from the compressed archive sent to the server'
+    task :augment_copy_exclude do
+      set :copy_exclude, (fetch(:copy_exclude) || []) + %w[node_modules tmp/*]
+    end
   end
 
+  before 'deploy:update_code', 'ndr_dev_support:augment_copy_exclude'
   before 'deploy:update_code', 'ndr_dev_support:remove_svn_cache_if_needed'
 end